Camera Wobbles & Vessel Sound disablement in IVA
Removed camera wobbles and vessel sounds when in IVA by modifying the game
settings directly and restoring them when going out of IVA or restarting
flight.

Known bugs :
-Game Settings don't recover when game crashes (but they do when Alt+F4,
the only problem is when the process gets killed brutally)
-Some sounds like the staging sound for example can't be disabled. I think
this is a bug from Squad. Funilly enough, these sounds' volume seems to
be proportional to atmosphere density around the vessel. PCR IVA buttons
seem to be affected by this.